+++ osu.Game/Overlays/OverlayPanelDisplayStyleControl.cs
@@ -11,6 +11,8 @@ using osu.Game.Graphics.UserInterface;
 using osu.Framework.Allocation;
 using osuTK.Graphics;
 using osu.Framework.Graphics.Cursor;
+using System.ComponentModel;
+using osu.Framework.Extensions;
 
 namespace osu.Game.Overlays
 {
@@ -52,7 +54,7 @@ namespace osu.Game.Overlays
             [Resolved]
             private OverlayColourProvider colourProvider { get; set; }
 
-            public string TooltipText => $@"{Value} 查看";
+            public string TooltipText => $@"{Value.GetDescription().ToString()}视图";
 
             private readonly SpriteIcon icon;
 
@@ -95,7 +97,9 @@ namespace osu.Game.Overlays
 
     public enum OverlayPanelDisplayStyle
     {
+        [Description("卡片")]
         Card,
+        [Description("列表")]
         List
     }
 }