+++ osu.Game/Overlays/SocialOverlay.cs
@@ -17,16 +17,15 @@ using osu.Game.Overlays.Social;
 using osu.Game.Users;
 using System.Threading;
 using osu.Framework.Allocation;
-using osu.Framework.Threading;
-using System.ComponentModel;
 using osu.Framework.Extensions.Color4Extensions;
+using osu.Framework.Threading;
 
 namespace osu.Game.Overlays
 {
     public class SocialOverlay : SearchableListOverlay<SocialTab, SocialSortCriteria, SortDirection>
     {
         private readonly LoadingSpinner loading;
-        private FillFlowContainer<SocialPanel> panels;
+        private FillFlowContainer<UserPanel> panels;
 
         protected override Color4 BackgroundColour => Color4Extensions.FromHex(@"60284b");
         protected override Color4 TrianglesColourLight => Color4Extensions.FromHex(@"672b51");
@@ -34,8 +33,6 @@ namespace osu.Game.Overlays
 
         protected override SearchableListHeader<SocialTab> CreateHeader() => new Header();
         protected override SearchableListFilterControl<SocialSortCriteria, SortDirection> CreateFilterControl() => new FilterControl();
-
-        protected override string LoginPlaceholder => @"查看在线玩家列表";
 
         private User[] users = Array.Empty<User>();
 
@@ -55,7 +52,7 @@ namespace osu.Game.Overlays
         }
 
         public SocialOverlay()
-             : base(OverlayColourScheme.Pink)
+            : base(OverlayColourScheme.Pink)
         {
             Add(loading = new LoadingSpinner());
 
@@ -123,7 +120,7 @@ namespace osu.Game.Overlays
             {
                 case SocialTab.Friends:
                     var friendRequest = new GetFriendsRequest(); // TODO filter arguments?
-                    friendRequest.Success += users => Users = users.Select(u => (User)u).ToArray();
+                    friendRequest.Success += users => Users = users.ToArray();
                     API.Queue(getUsersRequest = friendRequest);
                     break;
 
@@ -161,7 +158,7 @@ namespace osu.Game.Overlays
             if (Filter.DisplayStyleControl.Dropdown.Current.Value == SortDirection.Descending)
                 sortedUsers = sortedUsers.Reverse();
 
-            var newPanels = new FillFlowContainer<SocialPanel>
+            var newPanels = new FillFlowContainer<UserPanel>
             {
                 RelativeSizeAxes = Axes.X,
                 AutoSizeAxes = Axes.Y,
@@ -169,20 +166,21 @@ namespace osu.Game.Overlays
                 Margin = new MarginPadding { Top = 10 },
                 ChildrenEnumerable = sortedUsers.Select(u =>
                 {
-                    SocialPanel panel;
+                    UserPanel panel;
 
                     switch (Filter.DisplayStyleControl.DisplayStyle.Value)
                     {
                         case PanelDisplayStyle.Grid:
-                            panel = new SocialGridPanel(u)
+                            panel = new UserGridPanel(u)
                             {
                                 Anchor = Anchor.TopCentre,
-                                Origin = Anchor.TopCentre
+                                Origin = Anchor.TopCentre,
+                                Width = 290,
                             };
                             break;
 
                         default:
-                            panel = new SocialListPanel(u);
+                            panel = new UserListPanel(u);
                             break;
                     }
 
@@ -239,15 +237,12 @@ namespace osu.Game.Overlays
                     clearPanels();
                     break;
             }
-            base.APIStateChanged(api, state);
         }
     }
 
     public enum SortDirection
     {
-        [Description("正序")]
         Ascending,
-        [Description("倒序")]
         Descending
     }
 }